requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
celery-progress==0.1.2
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.13
django-braces==1.15.0
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-timezone-field==4.2.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
mpmath==1.2.1
msgpack==1.0.3
packaging==21.3
prompt-toolkit==3.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
zope.interface==5.4.0
